+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- Merge secrets on re-login and surface GraphQL errors
+
 ## [1.3.0] - 2026-08-24
 
 ### Added